# docs/contracts/poundcake-openapi.json
{
  "components": {
    "schemas": {
...
      "DeleteResponse": {
        "description": "Generic delete response.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "status": {
            "default": "deleted",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngredientCreate": {
        "description": "Schema for creating a new global ingredient.",
        "properties": {
          "action_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Deprecated alias for execution_id",
            "title": "Action Id"
          },
          "destination_target": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Destination Target"
          },
          "execution_engine": {
            "default": "undefined",
            "maxLength": 50,
            "title": "Execution Engine",
            "type": "string"
          },
          "execution_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Id"
          },
          "execution_parameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Parameters"
          },
          "execution_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Payload"
          },
          "execution_purpose": {
            "default": "utility",
            "enum": [
              "remediation",
              "comms",
              "utility"
            ],
            "title": "Execution Purpose",
            "type": "string"
          },
          "execution_target": {
            "maxLength": 100,
            "title": "Execution Target",
            "type": "string"
          },
          "expected_duration_sec": {
            "exclusiveMinimum": 0.0,
            "title": "Expected Duration Sec",
            "type": "integer"
          },
          "ingredient_kind": {
            "anyOf": [
              {
                "enum": [
                  "remediation",
                  "comms",
                  "utility"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Deprecated alias for execution_purpose",
            "title": "Ingredient Kind"
          },
          "is_blocking": {
            "default": true,
            "title": "Is Blocking",
            "type": "boolean"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "on_failure": {
            "default": "stop",
            "enum": [
              "continue",
              "stop",
              "retry"
            ],
            "title": "On Failure",
            "type": "string"
          },
          "retry_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Retry Count",
            "type": "integer"
          },
          "retry_delay": {
            "default": 5,
            "minimum": 0.0,
            "title": "Retry Delay",
            "type": "integer"
          },
          "task_key_template": {
            "maxLength": 255,
            "title": "Task Key Template",
            "type": "string"
          },
          "timeout_duration_sec": {
            "default": 300,
            "exclusiveMinimum": 0.0,
            "title": "Timeout Duration Sec",
            "type": "integer"
          }
        },
        "required": [
          "execution_target",
          "task_key_template",
          "expected_duration_sec"
        ],
        "title": "IngredientCreate",
        "type": "object"
      },
      "IngredientResponse": {
        "description": "Schema for ingredient responses (includes DB fields).",
        "properties": {
          "action_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Deprecated alias for execution_id",
            "title": "Action Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "destination_target": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Destination Target"
          },
          "execution_engine": {
            "default": "undefined",
            "maxLength": 50,
            "title": "Execution Engine",
            "type": "string"
          },
          "execution_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Id"
          },
          "execution_parameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Parameters"
          },
          "execution_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Payload"
          },
          "execution_purpose": {
            "default": "utility",
            "enum": [
              "remediation",
              "comms",
              "utility"
            ],
            "title": "Execution Purpose",
            "type": "string"
          },
          "execution_target": {
            "maxLength": 100,
            "title": "Execution Target",
            "type": "string"
          },
          "expected_duration_sec": {
            "exclusiveMinimum": 0.0,
            "title": "Expected Duration Sec",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ingredient_kind": {
            "anyOf": [
              {
                "enum": [
                  "remediation",
                  "comms",
                  "utility"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Deprecated alias for execution_purpose",
            "title": "Ingredient Kind"
          },
          "is_blocking": {
            "default": true,
            "title": "Is Blocking",
            "type": "boolean"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "on_failure": {
            "default": "stop",
            "enum": [
              "continue",
              "stop",
              "retry"
            ],
            "title": "On Failure",
            "type": "string"
          },
          "retry_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Retry Count",
            "type": "integer"
          },
          "retry_delay": {
            "default": 5,
            "minimum": 0.0,
            "title": "Retry Delay",
            "type": "integer"
          },
          "task_key_template": {
            "maxLength": 255,
            "title": "Task Key Template",
            "type": "string"
          },
          "timeout_duration_sec": {
            "default": 300,
            "exclusiveMinimum": 0.0,
            "title": "Timeout Duration Sec",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "execution_target",
          "task_key_template",
          "expected_duration_sec",
          "id",
          "created_at",
          "updated_at",
          "deleted"
        ],
        "title": "IngredientResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PoundCake API",
    "version": "2.0.131"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ingredients/": {
      "get": {
        "description": "List global ingredients with optional filtering.",
        "operationId": "list_ingredients_api_v1_ingredients__get",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngredientResponse"
                  },
                  "title": "Response List Ingredients Api V1 Ingredients  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Ingredients",
        "tags": [
          "logic"
        ]
      },
      "post": {
        "description": "Create a new global ingredient.",
        "operationId": "create_ingredient_api_v1_ingredients__post",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ingredient",
        "tags": [
          "logic"
        ]
      }
    },
    "/api/v1/ingredients/by-name/{recipe_name}": {
      "get": {
        "description": "List ingredients for a recipe name, ordered by step_order.",
        "operationId": "get_ingredients_by_recipe_name_api_v1_ingredients_by_name__recipe_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_name",
            "required": true,
            "schema": {
              "title": "Recipe Name",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngredientResponse"
                  },
                  "title": "Response Get Ingredients By Recipe Name Api V1 Ingredients By Name  Recipe Name  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ingredients By Recipe Name",
        "tags": [
          "logic"
        ]
      }
    },
    "/api/v1/ingredients/by-recipe/{recipe_id}": {
      "get": {
        "description": "List ingredients for a recipe ID, ordered by step_order.",
        "operationId": "get_ingredients_by_recipe_id_api_v1_ingredients_by_recipe__recipe_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "title": "Recipe Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngredientResponse"
                  },
                  "title": "Response Get Ingredients By Recipe Id Api V1 Ingredients By Recipe  Recipe Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ingredients By Recipe Id",
        "tags": [
          "logic"
        ]
      }
    },
    "/api/v1/ingredients/{ingredient_id}": {
      "delete": {
        "description": "Delete a global ingredient.",
        "operationId": "delete_ingredient_api_v1_ingredients__ingredient_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "ingredient_id",
            "required": true,
            "schema": {
              "title": "Ingredient Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Ingredient",
        "tags": [
          "logic"
        ]
      },
      "get": {
        "description": "Fetch a single ingredient by ID.",
        "operationId": "get_ingredient_api_v1_ingredients__ingredient_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ingredient_id",
            "required": true,
            "schema": {
              "title": "Ingredient Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ingredient",
        "tags": [
          "logic"
        ]
      },
      "patch": {
        "description": "Update a global ingredient.",
        "operationId": "update_ingredient_api_v1_ingredients__ingredient_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ingredient_id",
            "required": true,
            "schema": {
              "title": "Ingredient Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Ingredient",
        "tags": [
          "logic"
        ]
      },
      "put": {
        "description": "Update a global ingredient.",
        "operationId": "update_ingredient_api_v1_ingredients__ingredient_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "ingredient_id",
            "required": true,
            "schema": {
              "title": "Ingredient Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Ingredient",
        "tags": [
          "logic"
        ]
      }
    },
...
  }
}